Create predictor.py
