+++ likelihood/models/deep/__init__.py
@@ -1,2 +1,3 @@
 from .autoencoders import *
 from .gan import *
+from .predictor import GetInsights
