+++ likelihood/models/deep/autoencoders.py
@@ -1,33 +1,17 @@
-import base64
-import io
 import logging
 import os
-import random
-import warnings
 from functools import partial
 from shutil import rmtree
 
-import matplotlib
-import matplotlib.colors as mcolors
-import matplotlib.pyplot as plt
-import networkx as nx
 import numpy as np
 import pandas as pd
-from IPython.display import HTML, display
-from matplotlib import cm
-from matplotlib.colors import Normalize
-from pandas.plotting import radviz
 
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
 logging.getLogger("tensorflow").setLevel(logging.ERROR)
 
-
-from typing import List
-
 import keras_tuner
 import tensorflow as tf
 from pandas.core.frame import DataFrame
-from sklearn.manifold import TSNE
 from tensorflow.keras.layers import InputLayer
 from tensorflow.keras.regularizers import l2
 
@@ -895,661 +879,3 @@ def setup_model(
 
     best_hps = tuner.get_best_hyperparameters(1)[0].values
     return best_model, pd.DataFrame(best_hps, index=["Value"])
-
-
-class GetInsights:
-    """
-    A class to analyze the output of a neural network model, including visualizations
-    of the weights, t-SNE representation, and feature statistics.
-
-    Parameters
-    ----------
-    model : `AutoClassifier`
-        The trained model to analyze.
-    inputs : `np.ndarray`
-        The input data for analysis.
-    """
-
-    def __init__(self, model: AutoClassifier, inputs: np.ndarray) -> None:
-        """
-        Initializes the GetInsights class.
-
-        Parameters
-        ----------
-        model : `AutoClassifier`
-            The trained model to analyze.
-        inputs : `np.ndarray`
-            The input data for analysis.
-        """
-        self.inputs = inputs
-        self.model = model
-
-        self.encoder_layer = (
-            self.model.encoder.layers[1]
-            if isinstance(self.model.encoder.layers[0], InputLayer)
-            else self.model.encoder.layers[0]
-        )
-        self.decoder_layer = self.model.decoder.layers[0]
-
-        self.encoder_weights = self.encoder_layer.get_weights()[0]
-        self.decoder_weights = self.decoder_layer.get_weights()[0]
-
-        self.sorted_names = self._generate_sorted_color_names()
-
-    def _generate_sorted_color_names(self) -> list:
-        """
-        Generate sorted color names based on their HSV values.
-
-        Parameters
-        ----------
-        `None`
-
-        Returns
-        -------
-        `list` : Sorted color names.
-        """
-        colors = dict(mcolors.BASE_COLORS, **mcolors.CSS4_COLORS)
-        by_hsv = sorted(
-            (tuple(mcolors.rgb_to_hsv(mcolors.to_rgba(color)[:3])), name)
-            for name, color in colors.items()
-        )
-        sorted_names = [name for hsv, name in by_hsv if hsv[1] > 0.4 and hsv[2] >= 0.4]
-        random.shuffle(sorted_names)
-        return sorted_names
-
-    def render_html_report(
-        self,
-        frac: float = 0.2,
-        top_k: int = 5,
-        threshold_factor: float = 1.0,
-        max_rows: int = 5,
-        **kwargs,
-    ) -> None:
-        """
-        Generate and display an embedded HTML report in a Jupyter Notebook cell.
-
-        Parameters
-        ----------
-        frac : float, optional
-            Fraction of input data to use in predictor analyzer.
-        top_k : int, optional
-            Top K weights to highlight in encoder-decoder graph.
-        threshold_factor : float, optional
-            Factor to filter edge weights in encoder-decoder graph.
-        max_rows : int, optional
-            Maximum number of rows to display in the statistics DataFrame.
-        **kwargs : dict, optional
-            Additional keyword arguments for customization.
-        """
-
-        # Section: Predictor Analysis
-        display(HTML("<h2 style='margin-top:20px;'>📊 Predictor Analysis</h2>"))
-        display(
-            HTML(
-                "<p>This section visualizes how the model predicts the data. "
-                "You will see original inputs, reconstructed outputs, and analyses such as t-SNE "
-                "that reduce dimensionality to visualize latent space clustering.</p>"
-            )
-        )
-        stats_df = self.predictor_analyzer(frac=frac, **kwargs)
-
-        # Section: Encoder-Decoder Graph
-        display(HTML("<h2 style='margin-top:30px;'>🔁 Encoder-Decoder Graph</h2>"))
-        display(
-            HTML(
-                "<p>This visualization displays the connections between layers in the encoder and decoder. "
-                "Edges with the strongest weights are highlighted to emphasize influential features "
-                "in the model's transformation.</p>"
-            )
-        )
-        self.viz_encoder_decoder_graphs(threshold_factor=threshold_factor, top_k=top_k)
-
-        # Section: Statistical Summary
-        display(HTML("<h2 style='margin-top:30px;'>📈 Statistical Summary</h2>"))
-        display(
-            HTML(
-                "<p>This table summarizes feature statistics grouped by predicted classes, "
-                "including means, standard deviations, and modes, providing insight into "
-                "feature distributions across different classes.</p>"
-            )
-        )
-
-        # Limit number of rows displayed
-        if max_rows is not None and max_rows > 0:
-            stats_to_display = stats_df.head(max_rows)
-        else:
-            stats_to_display = stats_df
-
-        display(
-            stats_to_display.style.set_table_attributes(
-                "style='display:inline;border-collapse:collapse;'"
-            )
-            .set_caption("Feature Summary per Class")
-            .set_properties(
-                **{
-                    "border": "1px solid #ddd",
-                    "padding": "8px",
-                    "text-align": "center",
-                }
-            )
-        )
-
-        # Footer
-        display(
-            HTML(
-                "<p style='color: gray; margin-top:30px;'>Report generated with "
-                "<code>GetInsights</code> class. For detailed customization, extend "
-                "<code>render_html_report</code>.</p>"
-            )
-        )
-
-    def viz_encoder_decoder_graphs(self, threshold_factor=1.0, top_k=5):
-        """
-        Visualize encoder and decoder layers as directed graphs and show feature importances.
-        Highlights the top_k strongest weight connections.
-        """
-        use_labels = hasattr(self, "y_labels") and self.y_labels is not None
-        if use_labels:
-            assert (
-                len(self.y_labels) == self.encoder_weights.shape[0]
-            ), "Mismatch between y_labels and encoder input size"
-
-        def get_top_k_edges(weights, labels_src, labels_dst, k):
-            flat_weights = np.abs(weights.flatten())
-            indices = np.argpartition(flat_weights, -k)[-k:]
-            top_k_flat_indices = indices[np.argsort(-flat_weights[indices])]
-
-            top_k_edges = []
-            for flat_index in top_k_flat_indices:
-                i, j = np.unravel_index(flat_index, weights.shape)
-                src_label = (
-                    self.y_labels[i]
-                    if (labels_src == "Input" and use_labels)
-                    else f"{labels_src}_{i}"
-                )
-                dst_label = f"{labels_dst}_{j}"
-                weight = weights[i, j]
-                top_k_edges.append((src_label, dst_label, weight))
-            return top_k_edges
-
-        def create_graph(weights, labels_src, labels_dst):
-            if labels_src == "Input" and use_labels:
-                input_nodes = list(self.y_labels)
-            else:
-                input_nodes = [f"{labels_src}_{i}" for i in range(weights.shape[0])]
-            output_nodes = [f"{labels_dst}_{j}" for j in range(weights.shape[1])]
-
-            G = nx.DiGraph()
-            G.add_nodes_from(input_nodes + output_nodes)
-
-            abs_weights = np.abs(weights)
-            threshold = threshold_factor * np.mean(abs_weights)
-            top_k_edges = get_top_k_edges(weights, labels_src, labels_dst, top_k)
-            top_k_set = set((u, v) for u, v, _ in top_k_edges)
-
-            for i, src in enumerate(input_nodes):
-                for j, dst in enumerate(output_nodes):
-                    w = weights[i, j]
-                    if abs(w) > threshold:
-                        G.add_edge(src, dst, weight=w, highlight=(src, dst) in top_k_set)
-
-            return G, input_nodes, output_nodes
-
-        def layout_graph(nodes_left, nodes_right, x_offset):
-            pos = {}
-            max_len = max(len(nodes_left), len(nodes_right))
-            for i, node in enumerate(nodes_left):
-                pos[node] = (x_offset, -i)
-            for j, node in enumerate(nodes_right):
-                pos[node] = (x_offset + 1, -j)
-            return pos
-
-        def draw_graph(G, pos, node_colors, title, subplot_index):
-            ax = plt.subplot(1, 2, subplot_index)
-            weights = [abs(G[u][v]["weight"]) for u, v in G.edges()]
-            norm = Normalize(vmin=min(weights), vmax=max(weights))
-            cmap = cm.get_cmap("coolwarm")
-
-            edge_colors = [cmap(norm(abs(G[u][v]["weight"]))) for u, v in G.edges()]
-            edge_widths = [1.0 + 2.0 * norm(abs(G[u][v]["weight"])) for u, v in G.edges()]
-
-            nx.draw(
-                G,
-                pos,
-                with_labels=True,
-                node_color=node_colors,
-                node_size=1000,
-                font_size=9,
-                edge_color=edge_colors,
-                width=edge_widths,
-                arrows=True,
-                alpha=0.95,
-            )
-            ax.set_title(title, fontsize=13)
-
-        # Build graphs
-        G_enc, enc_inputs, enc_latents = create_graph(self.encoder_weights, "Input", "Latent")
-        G_dec, dec_latents, dec_outputs = create_graph(self.decoder_weights, "Latent", "Output")
-
-        pos_enc = layout_graph(enc_inputs, enc_latents, x_offset=0)
-        pos_dec = layout_graph(dec_latents, dec_outputs, x_offset=3)
-
-        plt.figure(figsize=(16, 6))
-        draw_graph(
-            G_enc,
-            pos_enc,
-            node_colors=["skyblue" if n in enc_inputs else "lightgreen" for n in G_enc.nodes()],
-            title="Encoder: Input → Latent",
-            subplot_index=1,
-        )
-        draw_graph(
-            G_dec,
-            pos_dec,
-            node_colors=["lightgreen" if n in dec_latents else "salmon" for n in G_dec.nodes()],
-            title="Decoder: Latent → Output",
-            subplot_index=2,
-        )
-
-        plt.suptitle(
-            f"Encoder-Decoder Graphs (Threshold: {threshold_factor:.2f}×Mean|w|, Top {top_k})",
-            fontsize=15,
-        )
-        plt.tight_layout(rect=[0, 0, 1, 0.95])
-        plt.show()
-
-        # Feature importance plot
-        importances = np.mean(np.abs(self.encoder_weights), axis=1)
-        sorted_idx = np.argsort(-importances)
-        xticks = [self.y_labels[i] if use_labels else f"Input_{i}" for i in sorted_idx]
-
-        plt.figure(figsize=(10, 4))
-        plt.bar(range(len(importances)), importances[sorted_idx], color="skyblue")
-        plt.xticks(range(len(importances)), xticks, rotation=45, ha="right")
-        plt.title("Feature Importances (Encoder Input Nodes)", fontsize=13)
-        plt.ylabel("Mean |Weight|")
-        plt.tight_layout()
-        plt.show()
-
-    def predictor_analyzer(
-        self,
-        frac: float = None,
-        cmap: str = "viridis",
-        aspect: str = "auto",
-        highlight: bool = True,
-        **kwargs,
-    ) -> None:
-        """
-        Analyze the model's predictions and visualize data.
-
-        Parameters
-        ----------
-        frac : `float`, optional
-            Fraction of data to use for analysis (default is `None`).
-        cmap : `str`, optional
-            The colormap for visualization (default is `"viridis"`).
-        aspect : `str`, optional
-            Aspect ratio for the visualization (default is `"auto"`).
-        highlight : `bool`, optional
-            Whether to highlight the maximum weights (default is `True`).
-        **kwargs : `dict`, optional
-            Additional keyword arguments for customization.
-
-        Returns
-        -------
-        `DataFrame` : The statistical summary of the input data.
-        """
-        self._viz_weights(cmap=cmap, aspect=aspect, highlight=highlight, **kwargs)
-        inputs = self.inputs.copy()
-        inputs = self._prepare_inputs(inputs, frac)
-        self.y_labels = kwargs.get("y_labels", None)
-        encoded, reconstructed = self._encode_decode(inputs)
-        self._visualize_data(inputs, reconstructed, cmap, aspect)
-        self._prepare_data_for_analysis(inputs, reconstructed, encoded, self.y_labels)
-
-        try:
-            self._get_tsne_repr(inputs, frac)
-            self._viz_tsne_repr(c=self.classification)
-
-            self._viz_radviz(self.data, "class", "Radviz Visualization of Latent Space")
-            self._viz_radviz(self.data_input, "class", "Radviz Visualization of Input Data")
-        except ValueError:
-            warnings.warn(
-                "Some functions or processes will not be executed for regression problems.",
-                UserWarning,
-            )
-
-        return self._statistics(self.data_input)
-
-    def _prepare_inputs(self, inputs: np.ndarray, frac: float) -> np.ndarray:
-        """
-        Prepare the input data, possibly selecting a fraction of it.
-
-        Parameters
-        ----------
-        inputs : `np.ndarray`
-            The input data.
-        frac : `float`
-            Fraction of data to use.
-
-        Returns
-        -------
-        `np.ndarray` : The prepared input data.
-        """
-        if frac:
-            n = int(frac * self.inputs.shape[0])
-            indexes = np.random.choice(np.arange(inputs.shape[0]), n, replace=False)
-            inputs = inputs[indexes]
-        inputs[np.isnan(inputs)] = 0.0
-        return inputs
-
-    def _encode_decode(self, inputs: np.ndarray) -> tuple:
-        """
-        Perform encoding and decoding on the input data.
-
-        Parameters
-        ----------
-        inputs : `np.ndarray`
-            The input data.
-
-        Returns
-        -------
-        `tuple` : The encoded and reconstructed data.
-        """
-        try:
-            mean, log_var = self.model.encoder(inputs)
-            encoded = sampling(mean, log_var)
-        except:
-            encoded = self.model.encoder(inputs)
-        reconstructed = self.model.decoder(encoded)
-        return encoded, reconstructed
-
-    def _visualize_data(
-        self, inputs: np.ndarray, reconstructed: np.ndarray, cmap: str, aspect: str
-    ) -> None:
-        """
-        Visualize the original data and the reconstructed data.
-
-        Parameters
-        ----------
-        inputs : `np.ndarray`
-            The input data.
-        reconstructed : `np.ndarray`
-            The reconstructed data.
-        cmap : `str`
-            The colormap for visualization.
-        aspect : `str`
-            Aspect ratio for the visualization.
-
-        Returns
-        -------
-        `None`
-        """
-        ax = plt.subplot(1, 2, 1)
-        plt.imshow(inputs, cmap=cmap, aspect=aspect)
-        plt.colorbar()
-        plt.title("Original Data")
-
-        plt.subplot(1, 2, 2, sharex=ax, sharey=ax)
-        plt.imshow(reconstructed, cmap=cmap, aspect=aspect)
-        plt.colorbar()
-        plt.title("Decoder Layer Reconstruction")
-        plt.show()
-
-    def _prepare_data_for_analysis(
-        self,
-        inputs: np.ndarray,
-        reconstructed: np.ndarray,
-        encoded: np.ndarray,
-        y_labels: List[str],
-    ) -> None:
-        """
-        Prepare data for statistical analysis.
-
-        Parameters
-        ----------
-        inputs : `np.ndarray`
-            The input data.
-        reconstructed : `np.ndarray`
-            The reconstructed data.
-        encoded : `np.ndarray`
-            The encoded data.
-        y_labels : `List[str]`
-            The labels of features.
-
-        Returns
-        -------
-        `None`
-        """
-        self.classification = (
-            self.model.classifier(tf.concat([reconstructed, encoded], axis=1))
-            .numpy()
-            .argmax(axis=1)
-        )
-
-        self.data = pd.DataFrame(encoded, columns=[f"Feature {i}" for i in range(encoded.shape[1])])
-        self.data_input = pd.DataFrame(
-            inputs,
-            columns=(
-                [f"Feature {i}" for i in range(inputs.shape[1])] if y_labels is None else y_labels
-            ),
-        )
-
-        self.data["class"] = self.classification
-        self.data_input["class"] = self.classification
-
-    def _get_tsne_repr(self, inputs: np.ndarray = None, frac: float = None) -> None:
-        """
-        Perform t-SNE dimensionality reduction on the input data.
-
-        Parameters
-        ----------
-        inputs : `np.ndarray`
-            The input data.
-        frac : `float`
-            Fraction of data to use.
-
-        Returns
-        -------
-        `None`
-        """
-        if inputs is None:
-            inputs = self.inputs.copy()
-            if frac:
-                n = int(frac * self.inputs.shape[0])
-                indexes = np.random.choice(np.arange(inputs.shape[0]), n, replace=False)
-                inputs = inputs[indexes]
-            inputs[np.isnan(inputs)] = 0.0
-        self.latent_representations = inputs @ self.encoder_weights
-
-        tsne = TSNE(n_components=2)
-        self.reduced_data_tsne = tsne.fit_transform(self.latent_representations)
-
-    def _viz_tsne_repr(self, **kwargs) -> None:
-        """
-        Visualize the t-SNE representation of the latent space.
-
-        Parameters
-        ----------
-        **kwargs : `dict`
-            Additional keyword arguments for customization.
-
-        Returns
-        -------
-        `None`
-        """
-        c = kwargs.get("c", None)
-        self.colors = (
-            kwargs.get("colors", self.sorted_names[: len(np.unique(c))]) if c is not None else None
-        )
-
-        plt.scatter(
-            self.reduced_data_tsne[:, 0],
-            self.reduced_data_tsne[:, 1],
-            cmap=matplotlib.colors.ListedColormap(self.colors) if c is not None else None,
-            c=c,
-        )
-
-        if c is not None:
-            cb = plt.colorbar()
-            loc = np.arange(0, max(c), max(c) / float(len(self.colors)))
-            cb.set_ticks(loc)
-            cb.set_ticklabels(np.unique(c))
-
-        plt.title("t-SNE Visualization of Latent Space")
-        plt.xlabel("t-SNE 1")
-        plt.ylabel("t-SNE 2")
-        plt.show()
-
-    def _viz_radviz(self, data: pd.DataFrame, color_column: str, title: str) -> None:
-        """
-        Visualize the data using RadViz.
-
-        Parameters
-        ----------
-        data : `pd.DataFrame`
-            The data to visualize.
-        color_column : `str`
-            The column to use for coloring.
-        title : `str`
-            The title of the plot.
-
-        Returns
-        -------
-        `None`
-        """
-        data_normalized = data.copy(deep=True)
-        data_normalized.iloc[:, :-1] = (
-            2.0
-            * (data_normalized.iloc[:, :-1] - data_normalized.iloc[:, :-1].min())
-            / (data_normalized.iloc[:, :-1].max() - data_normalized.iloc[:, :-1].min())
-            - 1
-        )
-        radviz(data_normalized, color_column, color=self.colors)
-        plt.title(title)
-        plt.show()
-
-    def _viz_weights(
-        self, cmap: str = "viridis", aspect: str = "auto", highlight: bool = True, **kwargs
-    ) -> None:
-        """
-        Visualize the encoder layer weights of the model.
-
-        Parameters
-        ----------
-        cmap : `str`, optional
-            The colormap for visualization (default is `"viridis"`).
-        aspect : `str`, optional
-            Aspect ratio for the visualization (default is `"auto"`).
-        highlight : `bool`, optional
-            Whether to highlight the maximum weights (default is `True`).
-        **kwargs : `dict`, optional
-            Additional keyword arguments for customization.
-
-        Returns
-        -------
-        `None`
-        """
-        title = kwargs.get("title", "Encoder Layer Weights (Dense Layer)")
-        y_labels = kwargs.get("y_labels", None)
-        cmap_highlight = kwargs.get("cmap_highlight", "Pastel1")
-        highlight_mask = np.zeros_like(self.encoder_weights, dtype=bool)
-
-        plt.imshow(self.encoder_weights, cmap=cmap, aspect=aspect)
-        plt.colorbar()
-        plt.title(title)
-        if y_labels is not None:
-            plt.yticks(ticks=np.arange(self.encoder_weights.shape[0]), labels=y_labels)
-        if highlight:
-            for i, j in enumerate(self.encoder_weights.argmax(axis=1)):
-                highlight_mask[i, j] = True
-            plt.imshow(
-                np.ma.masked_where(~highlight_mask, self.encoder_weights),
-                cmap=cmap_highlight,
-                alpha=0.5,
-                aspect=aspect,
-            )
-        plt.show()
-
-    def _statistics(self, data_input: DataFrame) -> DataFrame:
-        """
-        Compute statistical summaries of the input data.
-
-        Parameters
-        ----------
-        data_input : `DataFrame`
-            The data to compute statistics for.
-
-        Returns
-        -------
-        `DataFrame` : The statistical summary of the input data.
-        """
-        data = data_input.copy(deep=True)
-
-        if not pd.api.types.is_string_dtype(data["class"]):
-            data["class"] = data["class"].astype(str)
-
-        data.ffill(inplace=True)
-        grouped_data = data.groupby("class")
-
-        numerical_stats = grouped_data.agg(["mean", "min", "max", "std", "median"])
-        numerical_stats.columns = ["_".join(col).strip() for col in numerical_stats.columns.values]
-
-        def get_mode(x):
-            mode_series = x.mode()
-            return mode_series.iloc[0] if not mode_series.empty else None
-
-        mode_stats = grouped_data.apply(get_mode, include_groups=False)
-        mode_stats.columns = [f"{col}_mode" for col in mode_stats.columns]
-        combined_stats = pd.concat([numerical_stats, mode_stats], axis=1)
-
-        return combined_stats.T
-
-
-########################################################################################
-
-if __name__ == "__main__":
-    # Example usage
-    import pandas as pd
-    from sklearn.datasets import load_iris
-    from sklearn.preprocessing import OneHotEncoder
-
-    # Load the dataset
-    iris = load_iris()
-
-    # Convert to a DataFrame for easy exploration
-    iris_df = pd.DataFrame(data=iris.data, columns=iris.feature_names)
-    iris_df["species"] = iris.target
-
-    X = iris_df.drop(columns="species")
-    y_labels = X.columns
-    X = X.values
-    y = iris_df["species"].values
-
-    X = np.asarray(X).astype(np.float32)
-
-    encoder = OneHotEncoder()
-    y = encoder.fit_transform(y.reshape(-1, 1)).toarray()
-    y = np.asarray(y).astype(np.float32)
-
-    model = AutoClassifier(
-        input_shape_parm=X.shape[1],
-        num_classes=3,
-        units=27,
-        activation="tanh",
-        num_layers=2,
-        dropout=0.2,
-    )
-    model.compile(
-        optimizer="adam",
-        loss=tf.keras.losses.CategoricalCrossentropy(),
-        metrics=[tf.keras.metrics.F1Score(threshold=0.5)],
-    )
-    model.fit(X, y, epochs=50, validation_split=0.2)
-
-    insights = GetInsights(model, X)
-    summary = insights.predictor_analyzer(frac=1.0, y_labels=y_labels)
-    insights._get_tsne_repr()
-    insights._viz_tsne_repr()
-    insights._viz_tsne_repr(c=iris_df["species"])
-    insights._viz_weights()
-    print(summary)
